Env: the rail correction pushed the wrong way when the bias was railed
A UBSan build found five undefined operations in the render path. Four were a
wrap the code intends; one was a live bug.

THE BUG. OffsetForChiffAmplitude takes the mean as nominal + bias, and both are
Q30 levels, so their sum passes INT32_MAX. MEASURED at a full-range note with
the bias at a rail: 1093729386 + 1073709056 = 2167438442, which is +2.02 and
sits ABOVE mean_max, so the mean should be pushed DOWN. It wrapped to
-2127528854, took the `< mean_min` branch, and was pushed UP instead. The
comparison is in Q29 now -- halving both first is what the stage's own bias
slope already does -- and the answer is formed unsigned, because it too can pass
INT32_MAX and only ever lands in the modular accumulator.

The other four are wraps that cancel, made defined where they happen: the bias
ramp (`bias_q31 + slope * run_samples` lands on bias_target, which is in range;
only the step wraps), its Q30 twin, and `chiff_slew_state_q26 << 4` on a
negative state -- the C reference cast the shifted result, which is too late.
Also `int16_t << 16` where voice.cc and oscillator.cc form the bias, and the
same in the host driver's bias rig.

build.sh now builds a second binary with -fsanitize=signed-integer-overflow,shift
and runs the extreme cases through it before the battery. Nothing else in the
suite can see this class: a signed overflow renders whatever the compiler felt
like and every check still passes.

golden re-recorded: 4 of 14 vectors move, by ONE LSB on ONE sample in 333056.
That is the half LSB the Q29 comparison costs on a DC offset, and it is the
whole behavioural cost. flash 108220 -> 108256, ram unchanged. Battery green on
8 seeds, anomaly 140 cases.

+++ tools/hosttest/driver.cc
@@ -110,14 +110,14 @@ static void RenderMs(double ms) {
   size_t n = (size_t)(ms * 45.0);
   for (size_t i = 0; i < n; i += kAudioBlockSize) {
     int16_t buffer[kAudioBlockSize];
-    int32_t bias_target_q31 = g_tremolo
-        ? static_cast<int32_t>(env.tremolo(g_tremolo)) << 16 : 0;
+    uint32_t bias_target_q31 = g_tremolo
+        ? static_cast<uint32_t>(env.tremolo(g_tremolo)) << 16 : 0;
     if (g_bias_lfo) {
       const bool high = ((g_block_counter / g_bias_lfo_blocks) & 1) == 0;
-      bias_target_q31 += (high ? g_bias_lfo : -g_bias_lfo) << 16;
+      bias_target_q31 += static_cast<uint32_t>(high ? g_bias_lfo : -g_bias_lfo) << 16;
     }
     ++g_block_counter;
-    env.RenderSamples(buffer, bias_target_q31);
+    env.RenderSamples(buffer, static_cast<int32_t>(bias_target_q31));
     if (g_chiff_trace) {
       // The amount, not the drive: drive, slew time and input are all pure
       // functions of it, and it costs no member to expose.
